requirements.txt
alembic==1.0.11
appdirs==1.4.3
asn1crypto==0.24.0
attrs==18.2.0
autobahn==19.1.1
Automat==0.7.0
backcall==0.1.0
black==19.3b0
bleach==3.1.0
certifi==2019.6.16
cffi==1.11.5
chardet==3.0.4
Click==7.0
coloredlogs==10.0
colorlog==4.0.2
constantly==15.1.0
cryptography==2.4.2
cycler==0.10.0
decorator==4.4.0
defusedxml==0.6.0
docopt==0.6.2
entrypoints==0.3
Flask==1.0.2
Flask-API==1.1
Flask-Cors==3.0.8
Flask-Migrate==2.5.2
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
gunicorn==19.9.0
humanfriendly==4.18
hyperlink==18.0.0
idna==2.8
incremental==17.5.0
ipykernel==5.1.1
ipython==7.6.1
ipython-genutils==0.2.0
ipywidgets==7.5.0
itsdangerous==1.1.0
jedi==0.14.1
Jinja2==2.10.1
jsonschema==3.0.1
jupyter==1.0.0
jupyter-client==5.3.1
jupyter-console==6.0.0
jupyter-core==4.5.0
kiwisolver==1.1.0
kraken-wsclient-py==0.0.4
krakenex==2.1.0
Mako==1.0.13
MarkupSafe==1.1.1
matplotlib==3.1.1
mistune==0.8.4
mpl-finance==0.10.0
nbconvert==5.5.0
nbformat==4.4.0
nose==1.3.7
notebook==6.0.0
numpy==1.16.4
pandas==0.24.2
pandocfilters==1.4.2
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
prometheus-client==0.7.1
prompt-toolkit==2.0.9
psycopg2==2.8.3
ptyprocess==0.6.0
pyasn1==0.4.5
pyasn1-modules==0.2.3
pycparser==2.19
Pygments==2.4.2
PyHamcrest==1.9.0
PyJWT==1.7.1
pyOpenSSL==18.0.0
pyparsing==2.4.0
pyrsistent==0.15.3
python-dateutil==2.8.0
python-editor==1.0.4
pytz==2019.1
pyzmq==18.0.2
qtconsole==4.5.1
requests==2.22.0
Send2Trash==1.5.0
service-identity==18.1.0
simplejson==3.16.0
six==1.12.0
SQLAlchemy==1.3.5
terminado==0.8.2
testpath==0.4.2
toml==0.10.0
tornado==6.0.3
traitlets==4.3.2
Twisted==18.9.0
txaio==18.8.1
urllib3==1.25.3
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.15.4
widgetsnbextension==3.5.0
zope.interface==4.6.0
